confirmDelete Ask a confirmation before suppression of a training + add snackBar after add of suppression + fixing maxId selector

+++ src/app/modules/training-list/pages/training-list/training-list.page.ts
@@ -1,43 +1,34 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
+import { Observable, of, Subject } from 'rxjs';
 
 import { TrainingModel } from 'src/app/core/models/training.model';
-import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { TrainingService } from 'src/app/core/services/training-list.service';
-import { Observable } from 'rxjs';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmActionComponent } from 'src/app/shared/components/confirm-action/confirm-action.component';
+import { TranslateService } from '@ngx-translate/core';
 
 @Component({
   selector: 'app-training-list',
   templateUrl: 'training-list.page.html',
   styleUrls: ['training-list.page.scss'],
 })
-export class TrainingListPage implements OnInit {
+export class TrainingListPage {
 
   trainingList$: Observable<TrainingModel[]>;
+  deleteConfirm$ = new Subject();
 
   newTrainingCreating = false;
-  newTrainingForm: FormGroup;
-
+  
   newTrainingValue: string;
 
   constructor(
-    private formBuilder: FormBuilder,
-    private trainingService: TrainingService
+    private trainingService: TrainingService,
+    private snackBar: MatSnackBar,
+    private dialog: MatDialog,
+    private translate: TranslateService
   ) {
-    this.initForm();
-      this.trainingList$ = this.trainingService.getTrainingList();
-  }
-
-  ngOnInit() {
-
-  }
-
-  /**
- * Initialise le formulaire
- */
-  initForm() {
-    this.newTrainingForm = this.formBuilder.group({
-      titleControl: ['', Validators.required]
-    });
+    this.trainingList$ = this.trainingService.getTrainingList();
   }
 
   openAddTraining() {
@@ -61,6 +52,29 @@ export class TrainingListPage implements OnInit {
   }
 
   deleteTraining(trainingId: number){
-    this.trainingService.deleteTraining(trainingId);
+    this.deleteConfirm$.next(false);
+    this.openDialog();    
+    this.deleteConfirm$.subscribe(
+      data => {
+        if(data){
+          this.trainingService.deleteTraining(trainingId);
+        }
+      }
+    )    
   }
+
+  openDialog(): void {
+    const dialogRef = this.dialog.open(ConfirmActionComponent, {
+      width: '250px',
+      data: {title: '', text: this.translate.instant('CONFIRM.DELETE_TRAINING') }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if(result) {
+        console.log('close', result)
+        this.deleteConfirm$.next(true);
+      }
+    });
+  }
+
 }
